getSummary API added, startGame refactored
(normally try to avoid commits this big)

To simplify data management of active games for getSummary(), implementation was changed from HashMap to an ArrayList.
HashMap is faster, but ArrayList is easier to manipulate, which currently is more important.
As a result, the Game class now owns it's own id, with a getter for access.
It also gets a conviencence method for summarizing the current game score.

startGame() was refactored to simply take the team names directly as input.
Using consistent interfaces can be useful, but it added a lot of overhead.

README.md updated with more assumptions and better examples of how to use the library